Switched back to double for x,dx,ddx

+++ src/nr/nr.h
@@ -49,8 +49,8 @@ typedef Z_NR<IntegerT> Integer;
 
 /* Floating-point type inside the SVP/CVP solver */
 typedef double enumf;
-//typedef double enumxt;
-typedef int enumxt;
+typedef double enumxt;
+//typedef int enumxt;
 
 /**
  * return type
